- Created bean classes for each input/output page - renamed autoQuote to vehicleQuote - changed username to email in login/sign up and removed email field - changed dwelling type and heating type in home quote to radio buttons - refactored userProfile to userProfileResult & editUserProfil to userProfile to be consistant with other file names - changed all instances of "name" to firstName and Lastname
