+++ Website/nbproject/private/private.properties
@@ -1,7 +1,7 @@
-deploy.ant.properties.file=C:\\Users\\tasha\\AppData\\Roaming\\NetBeans\\8.2\\config\\GlassFishEE6\\Properties\\gfv3395937658.properties
+deploy.ant.properties.file=/home/rob/.netbeans/8.2/config/GlassFishEE6/Properties/gfv3-297679770.properties
 j2ee.platform.is.jsr109=true
-j2ee.server.domain=C:/Users/tasha/AppData/Roaming/NetBeans/8.2/config/GF_4.1.1/domain1
-j2ee.server.home=C:/Program Files/glassfish-4.1.1/glassfish
-j2ee.server.instance=[C:\\Program Files\\glassfish-4.1.1\\glassfish;C:\\Program Files\\glassfish-4.1.1\\glassfish\\domains\\domain1]deployer:gfv3ee6wc:localhost:4848
-j2ee.server.middleware=C:/Program Files/glassfish-4.1.1
-user.properties.file=C:\\Users\\tasha\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
+j2ee.server.domain=/home/rob/GlassFish_Server/glassfish/domains/domain1
+j2ee.server.home=/home/rob/GlassFish_Server/glassfish
+j2ee.server.instance=[/home/rob/GlassFish_Server/glassfish:/home/rob/GlassFish_Server/glassfish/domains/domain1]deployer:gfv3ee6wc:localhost:4848
+j2ee.server.middleware=/home/rob/GlassFish_Server
+user.properties.file=/home/rob/.netbeans/8.2/build.properties
