+++ Website/nbproject/private/private.xml
@@ -2,18 +2,6 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/D:/PR2460%20-%20Project/Website/web/option.xhtml</file>
-            <file>file:/D:/PR2460%20-%20Project/Website/web/signUp.xhtml</file>
-            <file>file:/D:/PR2460%20-%20Project/Website/web/autoQuoteResult.xhtml</file>
-            <file>file:/D:/PR2460%20-%20Project/Website/web/autoQuote.xhtml</file>
-            <file>file:/D:/PR2460%20-%20Project/Website/web/editUserProfile.xhtml</file>
-            <file>file:/D:/PR2460%20-%20Project/Website/web/login.xhtml</file>
-            <file>file:/D:/PR2460%20-%20Project/Website/web/userProfile.xhtml</file>
-            <file>file:/D:/PR2460%20-%20Project/Website/web/secondDriverInfo.xhtml</file>
-            <file>file:/D:/PR2460%20-%20Project/Website/web/index.xhtml</file>
-            <file>file:/D:/PR2460%20-%20Project/Website/src/java/ValidationBean.java</file>
-            <file>file:/D:/PR2460%20-%20Project/Website/web/homeQuote.xhtml</file>
-        </group>
+        <group/>
     </open-files>
 </project-private>
