+++ Website/src/java/HomeQuoteBean.java
@@ -1,0 +1,115 @@
+import java.io.Serializable;
+import javax.faces.bean.ManagedBean;
+
+@ManagedBean(name = "HomeQuoteBean")
+public class ValidationBean implements Serializable{
+
+    private String homeYear;
+    private String homeType;
+    private String heatingType;
+    private String address;
+    private String city;
+    private String province;
+    private String postalCode;
+    private Date startDate;
+    private Date enddate;
+    private Double basePremium;
+    private Double tax;
+    private Double total;
+
+    public String getHomeYear() {
+        return homeYear;
+    }
+
+    public void setHomeYear(String homeYear) {
+        this.homeYear = homeYear;
+    }
+
+    public String getHomeType() {
+        return homeType;
+    }
+
+    public void setHomeType(String homeType) {
+        this.homeType = homeType;
+    }
+
+    public String getHeatingType() {
+        return heatingType;
+    }
+
+    public void setHeatingType(String heatingType) {
+        this.heatingType = heatingType;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public String getCity() {
+        return city;
+    }
+
+    public void setCity(String city) {
+        this.city = city;
+    }
+
+    public String getProvince() {
+        return province;
+    }
+
+    public void setProvince(String province) {
+        this.province = province;
+    }
+
+    public String getPostalCode() {
+        return postalCode;
+    }
+
+    public void setPostalCode(String postalCode) {
+        this.postalCode = postalCode;
+    }
+
+    public Date getStartDate() {
+        return startDate;
+    }
+
+    public void setStartDate(Date startDate) {
+        this.startDate = startDate;
+    }
+
+    public Date getEnddate() {
+        return enddate;
+    }
+
+    public void setEnddate(Date enddate) {
+        this.enddate = enddate;
+    }
+
+    public Double getBasePremium() {
+        return basePremium;
+    }
+
+    public void setBasePremium(Double basePremium) {
+        this.basePremium = basePremium;
+    }
+
+    public Double getTax() {
+        return tax;
+    }
+
+    public void setTax(Double tax) {
+        this.tax = tax;
+    }
+
+    public Double getTotal() {
+        return total;
+    }
+
+    public void setTotal(Double total) {
+        this.total = total;
+    }
+}
