--- Website/src/java/ValidationBean.java
@@ -1,295 +1,0 @@
-import java.io.Serializable;
-import javax.faces.bean.ManagedBean;
-
-@ManagedBean(name = "validationBean")
-public class ValidationBean implements Serializable{
-
-    private String name;
-    private String gender;
-    private String dob;
-    private String phone;
-    private String email;
-    private String address;
-    private String city;
-    private String province;
-    private String postalCode;
-    private String licenseNumber;
-    private String licenseDate;
-    private String name2;
-    private String gender2;
-    private String dob2;
-    private String phone2;
-    private String email2;
-    private String address2;
-    private String city2;
-    private String province2;
-    private String postalCode2;
-    private String licenseNumber2;
-    private String licenseDate2;
-    private String username;
-    private String password;
-    private String secondPassword;
-    private String vehicleModel;
-    private String vehicleMake;
-    private String vehicleYear;
-    private String plateNumber;
-    private String homeYear;
-    private String homeType;
-    private String heatingType;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getGender() {
-        return gender;
-    }
-
-    public void setGender(String gender) {
-        this.gender = gender;
-    }
-
-    public String getDob() {
-        return dob;
-    }
-
-    public void setDob(String dob) {
-        this.dob = dob;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public String getProvince() {
-        return province;
-    }
-
-    public void setProvince(String province) {
-        this.province = province;
-    }
-
-    public String getPostalCode() {
-        return postalCode;
-    }
-
-    public void setPostalCode(String postalCode) {
-        this.postalCode = postalCode;
-    }
-
-    public String getLicenseNumber() {
-        return licenseNumber;
-    }
-
-    public void setLicenseNumber(String licenseNumber) {
-        this.licenseNumber = licenseNumber;
-    }
-
-    public String getLicenseDate() {
-        return licenseDate;
-    }
-
-    public void setLicenseDate(String licenseDate) {
-        this.licenseDate = licenseDate;
-    }
-
-    public String getName2() {
-        return name2;
-    }
-
-    public void setName2(String name2) {
-        this.name2 = name2;
-    }
-
-    public String getGender2() {
-        return gender2;
-    }
-
-    public void setGender2(String gender2) {
-        this.gender2 = gender2;
-    }
-
-    public String getDob2() {
-        return dob2;
-    }
-
-    public void setDob2(String dob2) {
-        this.dob2 = dob2;
-    }
-
-    public String getPhone2() {
-        return phone2;
-    }
-
-    public void setPhone2(String phone2) {
-        this.phone2 = phone2;
-    }
-
-    public String getEmail2() {
-        return email2;
-    }
-
-    public void setEmail2(String email2) {
-        this.email2 = email2;
-    }
-
-    public String getAddress2() {
-        return address2;
-    }
-
-    public void setAddress2(String address2) {
-        this.address2 = address2;
-    }
-
-    public String getCity2() {
-        return city2;
-    }
-
-    public void setCity2(String city2) {
-        this.city2 = city2;
-    }
-
-    public String getProvince2() {
-        return province2;
-    }
-
-    public void setProvince2(String province2) {
-        this.province2 = province2;
-    }
-
-    public String getPostalCode2() {
-        return postalCode2;
-    }
-
-    public void setPostalCode2(String postalCode2) {
-        this.postalCode2 = postalCode2;
-    }
-
-    public String getLicenseNumber2() {
-        return licenseNumber2;
-    }
-
-    public void setLicenseNumber2(String licenseNumber2) {
-        this.licenseNumber2 = licenseNumber2;
-    }
-
-    public String getLicenseDate2() {
-        return licenseDate2;
-    }
-
-    public void setLicenseDate2(String licenseDate2) {
-        this.licenseDate2 = licenseDate2;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getSecondPassword() {
-        return secondPassword;
-    }
-
-    public void setSecondPassword(String secondPassword) {
-        this.secondPassword = secondPassword;
-    }
-
-    public String getVehicleModel() {
-        return vehicleModel;
-    }
-
-    public void setVehicleModel(String vehicleModel) {
-        this.vehicleModel = vehicleModel;
-    }
-
-    public String getVehicleMake() {
-        return vehicleMake;
-    }
-
-    public void setVehicleMake(String vehicleMake) {
-        this.vehicleMake = vehicleMake;
-    }
-
-    public String getVehicleYear() {
-        return vehicleYear;
-    }
-
-    public void setVehicleYear(String vehicleYear) {
-        this.vehicleYear = vehicleYear;
-    }
-
-    public String getPlateNumber() {
-        return plateNumber;
-    }
-
-    public void setPlateNumber(String plateNumber) {
-        this.plateNumber = plateNumber;
-    }
-
-    public String getHomeYear() {
-        return homeYear;
-    }
-
-    public void setHomeYear(String homeYear) {
-        this.homeYear = homeYear;
-    }
-
-    public String getHomeType() {
-        return homeType;
-    }
-
-    public void setHomeType(String homeType) {
-        this.homeType = homeType;
-    }
-
-    public String getHeatingType() {
-        return heatingType;
-    }
-
-    public void setHeatingType(String heatingType) {
-        this.heatingType = heatingType;
-    }
-}
