+++ Website/src/java/VehicleQuoteBean.java
@@ -1,0 +1,98 @@
+import java.io.Serializable;
+import javax.faces.bean.ManagedBean;
+
+@ManagedBean(name = "VehicleQuoteBean")
+public class ValidationBean implements Serializable{
+
+    private Double vehicleValue;
+    private String vehicleModel;
+    private String vehicleMake;
+    private String vehicleYear;
+    private String plateNumber;
+    private Date startDate;
+    private Date enddate;
+    private Double basePremium;
+    private Double tax;
+    private Double total;
+
+    public Double getVehicleValue() {
+        return vehicleValue;
+    }
+
+    public void setVehicleValue(Double vehicleValue) {
+        this.vehicleValue = vehicleValue;
+    }
+
+    public String getVehicleModel() {
+        return vehicleModel;
+    }
+
+    public void setVehicleModel(String vehicleModel) {
+        this.vehicleModel = vehicleModel;
+    }
+
+    public String getVehicleMake() {
+        return vehicleMake;
+    }
+
+    public void setVehicleMake(String vehicleMake) {
+        this.vehicleMake = vehicleMake;
+    }
+
+    public String getVehicleYear() {
+        return vehicleYear;
+    }
+
+    public void setVehicleYear(String vehicleYear) {
+        this.vehicleYear = vehicleYear;
+    }
+
+    public String getPlateNumber() {
+        return plateNumber;
+    }
+
+    public void setPlateNumber(String plateNumber) {
+        this.plateNumber = plateNumber;
+    }
+
+    public Date getStartDate() {
+        return startDate;
+    }
+
+    public void setStartDate(Date startDate) {
+        this.startDate = startDate;
+    }
+
+    public Date getEnddate() {
+        return enddate;
+    }
+
+    public void setEnddate(Date enddate) {
+        this.enddate = enddate;
+    }
+
+    public Double getBasePremium() {
+        return basePremium;
+    }
+
+    public void setBasePremium(Double basePremium) {
+        this.basePremium = basePremium;
+    }
+
+    public Double getTax() {
+        return tax;
+    }
+
+    public void setTax(Double tax) {
+        this.tax = tax;
+    }
+
+    public Double getTotal() {
+        return total;
+    }
+
+    public void setTotal(Double total) {
+        this.total = total;
+    }
+
+}
