+++ Website/web/homeQuote.xhtml
@@ -17,49 +17,54 @@
              <h:outputLabel>Year Built:</h:outputLabel>
             <h:inputText id="homeYearInputText" required="true"
                 requiredMessage="Please enter year built(yyyy)"
-                value="#{validationBean.homeYear}"
+                value="#{HomeQuoteBean.homeYear}"
                 validatorMessage="Invalid year format">
                 <f:validateRegex pattern="^\d{4}$"/>
             </h:inputText>
             <h:outputLabel>Type of Dwelling:</h:outputLabel>
-            <h:inputText id="homeTypeInputText" required="true"
-                requiredMessage="Please enter type of dwelling"
-                value="#{validationBean.homeType}"
-                validatorMessage="Type of dwelling must be fewer than 20 characters">
-                <f:validateLength maximum="20"/>
-            </h:inputText>
+            <h:selectOneRadio id="dwellingRadio" required="true"
+                              requiredMessage="Please enter type of dwelling"
+                              value="#{HomeQuoteBean.homeType}">
+               <f:selectItem itemValue="Single Dwelling" itemLabel="Single Dwelling"/>
+               <f:selectItem itemValue="Apartment" itemLabel="Apartment"/>
+               <f:selectItem itemValue="Bungalow" itemLabel="Bungalow"/>
+               <f:selectItem itemValue="Semi-Attached" itemLabel="Semi-Attached"/>
+            </h:selectOneRadio>
             <h:outputLabel>Heating Type:</h:outputLabel>
-            <h:inputText id="heatingTypeInputText" required="true"
-                requiredMessage="Please enter heating type"
-                value="#{validationBean.heatingType}"
-                validatorMessage="Heating type must be fewer than 25 characters">
-                <f:validateLength maximum="25"/>
-            </h:inputText>
+            <h:selectOneRadio id="heatingRadio" required="true"
+                              requiredMessage="Please enter type of heating"
+                              value="#{HomeQuoteBean.heatingType}">
+               <f:selectItem itemValue="Electric" itemLabel="Electric"/>
+               <f:selectItem itemValue="Oil" itemLabel="Oil"/>
+               <f:selectItem itemValue="Wood" itemLabel="Wood"/>
+               <f:selectItem itemValue="Gas" itemLabel="Gas"/>
+               <f:selectItem itemValue="Other" itemLabel="Other"/>
+            </h:selectOneRadio>
             <h:outputLabel>Address:</h:outputLabel>
             <h:inputText id="addressInputText" required="true"
                 requiredMessage="Please enter address"
-                value="#{validationBean.address}"
+                value="#{HomeQuoteBean.address}"
                 validatorMessage="Address must be fewer than 30 characters">
                 <f:validateLength maximum="30"/>
             </h:inputText>
             <h:outputLabel>City:</h:outputLabel>
             <h:inputText id="cityInputText" required="true"
                 requiredMessage="Please enter city"
-                value="#{validationBean.city}"
+                value="#{HomeQuoteBean.city}"
                 validatorMessage="City must be fewer than 15 characters">
                 <f:validateLength maximum="15"/>
             </h:inputText>
             <h:outputLabel>Province:</h:outputLabel>
             <h:inputText id="provinceInputText" required="true"
                 requiredMessage="Please enter province"
-                value="#{validationBean.province}"
+                value="#{HomeQuoteBean.province}"
                 validatorMessage="Province must be at least than 2 characters">
                 <f:validateLength minimum="2"/>
             </h:inputText>
             <h:outputLabel>Postal Code:</h:outputLabel>
             <h:inputText id="postalCodeInputText" required="true"
                 requiredMessage="Please enter postal code"
-                value="#{validationBean.postalCode}"
+                value="#{HomeQuoteBean.postalCode}"
                 validatorMessage="Invalid postal code format">
                 <f:validateRegex pattern="^(?!.*[DFIOQU])[A-VXY][0-9][A-Z] ?[0-9][A-Z][0-9]$"/>
             </h:inputText>
