+++ Website/web/secondDriverInfo.xhtml
@@ -14,81 +14,88 @@
          <h1>Secondary Driver Information</h1>
          <p>Please fill in all fields and click submit</p>
          <h:panelGrid columns="4" style="height: 96px; width:456px;">
-            <h:outputLabel>Name:</h:outputLabel>
-            <h:inputText id="name2InputText" required="true"
-                requiredMessage="Please enter name"
-                value="#{validationBean.name2}"
+            <h:outputLabel>First Name:</h:outputLabel>
+            <h:inputText id="firstNameInputText" required="true"
+                requiredMessage="Please enter first name"
+                value="#{SecondaryDriverBean.firstName}"
                 validatorMessage="First name must be fewer than 40 characters">
                 <f:validateLength maximum="40"/>
             </h:inputText>
+            <h:outputLabel>Last Name:</h:outputLabel>
+            <h:inputText id="lastNameInputText" required="true"
+                         requiredMessage="Please enter last name"
+                         value="#{SecondaryDriverBean.lastName}"
+                         validatorMessage="Last name must be fewer than 40 characters">
+               <f:validateLength maximum="40"/>
+            </h:inputText>
             <h:outputLabel>Gender:</h:outputLabel>
-            <h:selectOneRadio id="gender2SelectOneRadio" required="true"
+            <h:selectOneRadio id="genderSelectOneRadio" required="true"
                requiredMessage="Please select a gender"
-               value="#{validationBean.gender2}">
+               value="#{SecondaryDriverBean.gender}">
                <f:selectItem itemValue="Male" itemLabel="Male"/>
                <f:selectItem itemValue="Female" itemLabel="Female"/>
                <f:selectItem itemValue="Other" itemLabel="Other"/>
             </h:selectOneRadio>
             <h:outputLabel>Date of Birth:</h:outputLabel>
-            <h:inputText id="dob2InputText" required="true"
+            <h:inputText id="dobInputText" required="true"
                 requiredMessage="Please enter date of birth(yyyy-mm-dd)"
-                value="#{validationBean.dob2}"
+                value="#{SecondaryDriverBean.dob}"
                 validatorMessage="Invalid date of birth format">
                 <f:validateRegex pattern="^\d{4}\-(0?[1-9]|1[012])\-(0?[1-9]|[12][0-9]|3[01])$"/>
             </h:inputText>
             <h:outputLabel>Phone:</h:outputLabel>
-            <h:inputText id="phone2InputText" required="true"
+            <h:inputText id="phoneInputText" required="true"
                 requiredMessage="Please enter phone number"
-                value="#{validationBean.phone2}"
+                value="#{SecondaryDriverBean.phone}"
                 validatorMessage="Invalid phone number format">
                 <f:validateRegex pattern="((\(\d{3}\)?)|(\d{3}-))?\d{3}-\d{4}"/>
             </h:inputText>
             <h:outputLabel>E-mail:</h:outputLabel>
-            <h:inputText id="email2InputText" required="true"
+            <h:inputText id="emailInputText" required="true"
                 requiredMessage="Please enter e-mail"
-                value="#{validationBean.email2}"
+                value="#{SecondaryDriverBean.email}"
                 validatorMessage="Invalid e-mail address format">
                 <f:validateRegex pattern="\w+([-+.']\w+)*@\w+([-.]\w+)*\.\w+([-.]\w+)*"/>
             </h:inputText>
             <h:outputLabel>Address:</h:outputLabel>
-            <h:inputText id="address2InputText" required="true"
+            <h:inputText id="addressInputText" required="true"
                 requiredMessage="Please enter address"
-                value="#{validationBean.address2}"
+                value="#{SecondaryDriverBean.address}"
                 validatorMessage="Address must be fewer than 30 characters">
                 <f:validateLength maximum="30"/>
             </h:inputText>
             <h:outputLabel>City:</h:outputLabel>
-            <h:inputText id="city2InputText" required="true"
+            <h:inputText id="cityInputText" required="true"
                 requiredMessage="Please enter city"
-                value="#{validationBean.city2}"
+                value="#{SecondaryDriverBean.city}"
                 validatorMessage="City must be fewer than 15 characters">
                 <f:validateLength maximum="15"/>
             </h:inputText>
             <h:outputLabel>Province:</h:outputLabel>
-            <h:inputText id="province2InputText" required="true"
+            <h:inputText id="provinceInputText" required="true"
                 requiredMessage="Please enter province"
-                value="#{validationBean.province2}"
-                validatorMessage="Province must be at least than 2 characters">
-                <f:validateLength minimum="2"/>
+                value="#{SecondaryDriverBean.province}"
+                validatorMessage="Province must be at least than  characters">
+                <f:validateLength minimum=""/>
             </h:inputText>
             <h:outputLabel>Postal Code:</h:outputLabel>
-            <h:inputText id="postalCode2InputText" required="true"
+            <h:inputText id="postalCodeInputText" required="true"
                 requiredMessage="Please enter postal code"
-                value="#{validationBean.postalCode2}"
+                value="#{SecondaryDriverBean.postalCode}"
                 validatorMessage="Invalid postal code format">
                 <f:validateRegex pattern="^[A-VXY][0-9][A-Z] ?[0-9][A-Z][0-9]$"/>
             </h:inputText>
             <h:outputLabel>Drivers License Number:</h:outputLabel>
-            <h:inputText id="licenseNumber2InputText" required="true"
+            <h:inputText id="licenseNumberInputText" required="true"
                 requiredMessage="Please enter drivers license number"
-                value="#{validationBean.licenseNumber2}"
+                value="#{SecondaryDriverBean.licenseNumber}"
                 validatorMessage="Drivers license number must be at least than 10 characters">
                 <f:validateLength minimum="10"/>
             </h:inputText>
             <h:outputLabel>License Issue Date:</h:outputLabel>
-            <h:inputText id="licenseDate2InputText" required="true"
+            <h:inputText id="licenseDateInputText" required="true"
                 requiredMessage="Please enter license issue date(yyyy-mm-dd)"
-                value="#{validationBean.licenseDate2}"
+                value="#{SecondaryDriverBean.licenseDate}"
                 validatorMessage="Invalid date format">
                 <f:validateRegex pattern="^\d{4}\-(0?[1-9]|1[012])\-(0?[1-9]|[12][0-9]|3[01])$"/>
             </h:inputText>
