+++ Website/web/userProfileResult.xhtml
@@ -1,0 +1,33 @@
+<?xml version='1.0' encoding='UTF-8' ?>
+
+<!-- userProfileResult.xhtml -->
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
+   "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml"
+      xmlns:h="http://java.sun.com/jsf/html"
+      xmlns:f="http://java.sun.com/jsf/core">
+   <h:head>
+      <title>User Information</title>
+   </h:head>
+   <h:body>
+      <h:form>
+         <h1>User Information</h1>
+         <h:outputLabel>Client ID:</h:outputLabel>
+         <h:panelGrid columns="4" style="height: 96px; width:456px;">
+            <h:outputLabel>First Name: #{UserBean.firstName}</h:outputLabel>
+            <h:outputLabel>First Name: #{UserBean.lastName}</h:outputLabel>
+            <h:outputLabel>Gender: #{UserBean.gender}</h:outputLabel>
+            <h:outputLabel>Date of Birth: #{UserBean.dob}</h:outputLabel>
+            <h:outputLabel>Phone: #{UserBean.phone}</h:outputLabel>
+            <h:outputLabel>Address: #{UserBean.address}</h:outputLabel>
+            <h:outputLabel>City: #{UserBean.city}</h:outputLabel>
+            <h:outputLabel>Province: #{UserBean.province}</h:outputLabel>
+            <h:outputLabel>Postal Code: #{UserBean.postalCode}</h:outputLabel>
+            <h:outputLabel>Drivers License Number: #{UserBean.licenseNumber}</h:outputLabel>
+            <h:outputLabel>License Issue Date: #{UserBean.licenseDate}</h:outputLabel>
+         </h:panelGrid>
+         <h:commandButton value="Edit" action="editUserProfile?faces-redirect=true"/>
+         <h:commandButton value="Cancel" action="option?faces-redirect=true" immediate="true"/>
+      </h:form>
+   </h:body>
+</html>
