+++ Website/web/vehicleQuote.xhtml
@@ -1,50 +1,57 @@
 <?xml version='1.0' encoding='UTF-8' ?>
 
-<!-- autoQuote.xhtml -->
+<!-- vehicleQuote.xhtml -->
 <!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
    "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
 <html xmlns="http://www.w3.org/1999/xhtml"
       xmlns:h="http://java.sun.com/jsf/html"
       xmlns:f="http://java.sun.com/jsf/core">
    <h:head>
-      <title>Auto Quote Form</title>
+      <title>Vehicle Quote Form</title>
    </h:head>
    <h:body>
       <h:form>
-         <h1>Automotive Quote</h1>
+         <h1>Vehicle Quote</h1>
          <p>Please fill in all fields and click Submit</p>
          <h:panelGrid columns="4" style="height: 96px; width:456px;">
+            <h:outputLabel>Vehicle Value:</h:outputLabel>
+            <h:inputText id="vehicleValueInputText" required="true"
+                         requiredMessage="Please enter vehicle value"
+                         value="#{VehicleQuoteBean.vehicleValue}"
+                         validatorMessage="Vehicle value must be a number">
+               <f:validateRegex pattern="^[0-9]*$"/>
+            </h:inputText>
             <h:outputLabel>Vehicle Make:</h:outputLabel>
             <h:inputText id="vehicleMakeInputText" required="true"
                 requiredMessage="Please enter vehicle make"
-                value="#{validationBean.vehicleMake}"
+                value="#{VehicleQuoteBean.vehicleMake}"
                 validatorMessage="Vehicle make must be fewer than 20 characters">
                 <f:validateLength maximum="20"/>
             </h:inputText>
             <h:outputLabel>Vehicle Model:</h:outputLabel>
             <h:inputText id="vehicleModelInputText" required="true"
                 requiredMessage="Please enter vehicle model"
-                value="#{validationBean.vehicleModel}"
+                value="#{VehicleQuoteBean.vehicleModel}"
                 validatorMessage="Vehicle model must be fewer than 20 characters">
                 <f:validateLength maximum="20"/>
             </h:inputText>
             <h:outputLabel>Vehicle Year:</h:outputLabel>
             <h:inputText id="vehicleYearInputText" required="true"
                 requiredMessage="Please enter vehicle year(yyyy)"
-                value="#{validationBean.vehicleYear}"
+                value="#{VehicleQuoteBean.vehicleYear}"
                 validatorMessage="Invalid year format">
                 <f:validateRegex pattern="^\d{4}$"/>
             </h:inputText>
             <h:outputLabel>Plate Number:</h:outputLabel>
             <h:inputText id="plateNumberInputText" required="true"
                 requiredMessage="Please enter plate number"
-                value="#{validationBean.plateNumber}"
+                value="#{VehicleQuoteBean.plateNumber}"
                 validatorMessage="Invalid plate number format">
                 <f:validateRegex pattern="^[A-Z]{1,3}[0-9]{1,4}$"/>
             </h:inputText>
             <h:commandButton value="Add Second Driver" action="secondDriverInfo?faces-redirect=true"/>
          </h:panelGrid>    
-         <h:commandButton value="Submit" action="autoQuoteResult?faces-redirect=true"/>
+         <h:commandButton value="Submit" action="vehicleQuoteResult?faces-redirect=true"/>
          <h:commandButton value="Cancel" action="option?faces-redirect=true" immediate="true"/>
       </h:form>
    </h:body>
