+++ Website/web/vehicleQuoteResult.xhtml
@@ -1,0 +1,36 @@
+<?xml version='1.0' encoding='UTF-8' ?>
+
+<!-- vehicleQuoteResult.xhtml -->
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
+   "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml"
+      xmlns:h="http://java.sun.com/jsf/html"
+      xmlns:f="http://java.sun.com/jsf/core">
+   <h:head>
+      <title>Vehicle Quote</title>
+   </h:head>
+   <h:body>
+      <h:form>
+         <h1>Vehicle Quote</h1>
+         <h:outputLabel>Quote ID:</h:outputLabel>
+         <h:panelGrid columns="4" style="height: 96px; width:456px;">
+            <h:outputLabel>Date Quoted: #{VehicleQuoteBean.startDate}</h:outputLabel>
+            <h:outputLabel>Expiry Date: #{VehicleQuoteBean.endDate}</h:outputLabel>
+            <h:outputLabel>Base Premium: #{VehicleQuoteBean.basePremium}</h:outputLabel>
+            <h:outputLabel>Tax: #{VehicleQuoteBean.tax}</h:outputLabel>
+            <h:outputLabel>Total: #{VehicleQuoteBean.total}</h:outputLabel>
+         </h:panelGrid> 
+         <h:outputLabel>Vehicle ID:</h:outputLabel>
+         <h:panelGrid columns="4" style="height: 96px; width:456px;">
+            <h:outputLabel>Vehicle Value: #{VehicleQuoteBean.vehicleValue}</h:outputLabel>
+            <h:outputLabel>Vehicle Make: #{VehicleQuoteBean.vehicleMake}</h:outputLabel>
+            <h:outputLabel>Vehicle Model: #{VehicleQuoteBean.vehicleModel}</h:outputLabel>
+            <h:outputLabel>Vehicle Year: #{VehicleQuoteBean.vehicleYear}</h:outputLabel>
+            <h:outputLabel>Plate Number: #{VehicleQuoteBean.plateNumber}</h:outputLabel>
+            <h:outputLabel>Secondary Driver: #{SecondaryDriverBean.firstName} #{SecondaryDriverBean.lastName}</h:outputLabel>
+         </h:panelGrid> 
+         <h:commandButton value="Options" action="option?faces-redirect=true" immediate="true"/>
+         <h:commandButton value="Generate Policy"/>
+      </h:form>
+   </h:body>
+</html>
